refactor(product-list): restructure ProductList import path and create new ProductList component
