+++ src/features/product/list-product/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ProductList';
+export { default } from './ui/ProductList/ProductList';
